requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Babel==4.0.0
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
markdown2==2.4.10
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
WTForms==3.1.1
email-validator==2.1.0
requests==2.31.0
